+++ NEWS.md
@@ -1,3 +1,12 @@
+# metamer 0.2.0.9000
+
+## New features 
+
+* Added `fortify()` method to plot metamers directly with 'ggplot2'.
+
+* The `perturbation` argument in `metamerize()` can now be a vector of equal in 
+length to `change`
+
 # metamer 0.1.0
 
 * First release